fix: address the six remaining review threads on PR #3
Coordinate precision (models.py, 0001_baseline_schema.py, grid.py, cache.py)

  farms.latitude/longitude were Float while docs/03-data-model.md specifies
  NUMERIC(9,6) and weather_grid_cells already used it. Both the model and the
  baseline migration now use Numeric(9, 6). The baseline is introduced by this
  PR and has not shipped, so it is corrected in place rather than followed by a
  second migration that fixes a schema nobody has.

  That change forces grid.snap() to handle Decimal: a coordinate loaded from a
  Numeric column reaches it through the ingest-by-farm path, and the old
  implementation did `Decimal / float`, which raises TypeError. snap() now
  accepts either type and returns Decimal quantized to the columns' 6 places,
  so a snapped coordinate compares equal to a stored one by construction.
  ROUND_HALF_EVEN preserves the behaviour of the round() it replaces.

  On the reviewer's related point about the cell lookup comparing a Numeric
  column to a Python float: that does not reproduce. Probed directly against
  PostgreSQL 16, float and Decimal lookups both hit, because SQLAlchemy's
  Numeric bind processor converts through str() first. The lookup is now exact
  by construction rather than by relying on that, but no observable bug was
  fixed there. get_or_create_cell also handles IntegrityError on the insert,
  so losing the uq_grid_cell_latlon race reads the winner instead of raising.

create_all against Alembic-managed databases (main.py)

  It ran unconditionally, so starting against an empty PostgreSQL database
  created tables Alembic never stamped and a later `alembic upgrade head`
  failed on tables that already existed. Guarded to SQLite, which is what the
  comment always claimed it was for.

Test and documentation fixes

  test_money.py: `assert ... or True` always passed. Removed the padding; the
  assertion holds on its own, since Decimal lands on exactly 0.3 while binary
  float overshoots to 0.30000000000000004.

  docs/README.md: the status block said implementation had not started, which
  contradicted the file's own header. It now matches, and points at document 10
  as the as-built record.

Verified on PostgreSQL 16, not only SQLite: app import against an empty
database creates 0 tables, `alembic upgrade head` then runs clean and stamps
0001_baseline, farms.latitude/longitude report numeric(9,6), two neighbouring
pins share one grid cell, re-snapping a stored coordinate finds that same cell,
and evaluation settles 31 of 31 observations with one trigger and one payout
reused on the second call. Suite is 47 passing, up from 44.

`alembic check` reports one remaining drift, ix_weather_observations_id, from
index=True on that table's primary key. It predates this PR and is outside
these six threads, so it is left alone.

+++ backend/app/models.py
@@ -61,8 +61,11 @@ class Farm(Base):
     id = Column(Integer, primary_key=True, index=True)
     farmer_name = Column(String, nullable=False)
     location = Column(String, nullable=False)
-    latitude = Column(Float, nullable=True)
-    longitude = Column(Float, nullable=True)
+    # NUMERIC(9, 6), matching weather_grid_cells: a coordinate must compare
+    # equal to itself after a database round-trip, or the same pin can snap to
+    # two different cells. Float cannot promise that.
+    latitude = Column(Numeric(9, 6), nullable=True)
+    longitude = Column(Numeric(9, 6), nullable=True)
     crop = Column(String, nullable=False)
     area_acres = Column(Float, nullable=False)
     crop_stage = Column(String, nullable=True)
